feat: Refactor Kiro Constellation VSCode extension for modular architecture
- Implemented ConfigurationLoader for multi-path environment variable loading
- Created PolarisController for handling file save events and documentation processing
- Established CommandController for managing VSCode command registrations
- Simplified extension entry point in extension.ts to under 150 lines
- Preserved existing service instantiation patterns and error handling mechanisms
- Ensured all functionality remains identical post-refactoring through comprehensive testing
