--- tasks/prd-draft-jsdoc-output-cleanup.md
@@ -1,85 +1,0 @@
-# DRAFT PRD: JSDoc Generation Output Cleanup
-
-## 1. Overview
-
-This feature addresses critical output quality issues in the existing JSDoc generation functionality. The current `constellation.testDocstringGeneration` command produces duplicate logging and incorrectly formatted JSDoc comments that contain formatting artifacts from the LLM response. The primary goal is to ensure the debug command produces clean, usable JSDoc output that can be directly copied and used by developers, while eliminating redundant console logging that creates noise in the debug console.
-
-## 2. Assumptions Made
-
-- The issue is specifically with the test command output formatting, not the core JSDoc generation logic
-- The LLMService.generateDocstring() method is the primary source of duplicate logging
-- The raw API response from OpenRouter contains markdown code fences and nested comment syntax
-- The existing validateAndProcessJSDocResponse() method may not be adequately sanitizing the response
-- The current prompt engineering is working correctly and doesn't need modification
-- The target output format should be a clean JSDoc block ready for direct insertion into code
-- Only the debug console output needs to be fixed, not the underlying API integration
-- The existing error handling and fallback mechanisms should remain unchanged
-
-## 3. User Stories
-
-- As a developer using the debug command, I want to see the generated JSDoc printed only once so that my debug console remains clean and easy to read.
-- As a developer testing JSDoc generation, I want the output to be a clean, valid comment block free of markdown or nested comment syntax so that I can copy and use it directly.
-- As a developer debugging the extension, I want to be able to trace the JSDoc generation process without redundant logging so that I can efficiently identify issues.
-
-## 4. Functional Requirements
-
-**FR1. Eliminate Duplicate Logging**
-- **Acceptance Criteria:**
-  - Given the `constellation.testDocstringGeneration` command is executed, when the JSDoc generation completes, then the final JSDoc block is logged to the debug console exactly once
-  - Given the command execution, when reviewing the debug console output, then there are no duplicate JSDoc blocks with different prefixes
-  - Given the logging flow, when tracing through the code, then each unique log message appears only once
-
-**FR2. Sanitize JSDoc Output**
-- **Acceptance Criteria:**
-  - Given the LLM returns a response with markdown code fences (```typescript), when the response is processed, then the code fences are completely removed from the final output
-  - Given the LLM returns nested comment blocks, when the response is sanitized, then only the outer JSDoc structure remains
-  - Given any formatting artifacts in the LLM response, when the sanitization process runs, then the output contains only valid JSDoc syntax
-
-**FR3. Validate Output Format**
-- **Acceptance Criteria:**
-  - Given a sanitized JSDoc output, when format validation runs, then it confirms the output starts with `/**` and ends with `*/`
-  - Given the final JSDoc output, when copied from the debug console, then it can be directly pasted into a TypeScript file without syntax errors
-  - Given invalid format detection, when the response cannot be properly sanitized, then the system falls back to the existing fallback JSDoc generation
-
-**FR4. Maintain Existing Functionality**
-- **Acceptance Criteria:**
-  - Given the output cleanup implementation, when the command executes, then all existing error handling mechanisms continue to function
-  - Given API failures or network issues, when the fallback JSDoc generation triggers, then it produces clean output without duplication
-  - Given the changes to logging, when the command completes successfully, then the user still receives the appropriate success notification
-
-## 5. Out of Scope (Non-Goals)
-
-- Modifying the LLM model or provider (OpenRouter)
-- Changing the core prompt engineering or system prompts sent to the AI
-- Implementing automatic insertion of JSDoc into source files
-- Modifying the overall architecture or API integration patterns
-- Adding new JSDoc generation features or capabilities
-- Changing the command name or registration mechanism
-- Modifying the test function used for validation
-
-## 6. Technical Considerations (Optional)
-
-- The duplicate logging appears to occur in the `LLMService.generateDocstring()` method around lines 147-154
-- The response sanitization should be enhanced in the `validateAndProcessJSDocResponse()` method
-- Existing `fixJSDocFormat()` method may need extension to handle markdown artifacts
-- The extension.ts command handler logging should be reviewed to ensure no duplication there
-- Consider adding a dedicated sanitization step before the existing format validation
-- Maintain backward compatibility with existing error handling patterns
-- Ensure changes don't impact the test suite in `integration.test.ts`
-
-## 7. Success Metrics
-
-- Reduce debug console noise by eliminating duplicate JSDoc logging (target: 50% reduction in redundant log entries)
-- Achieve 100% clean JSDoc output format (no markdown artifacts or nested comments in final output)
-- Maintain 100% compatibility with existing error handling and fallback mechanisms
-
-## 8. Open Questions
-
-1. Should the intermediate logging steps (like "Generated JSDoc:" and "Final validated JSDoc:") be consolidated into a single final output log? Just makie the logging simple enouigh, that gives context.
-2. What specific markdown patterns should the sanitization process target beyond code fences (```typescript```)? We need to make it generic for code, we cna keep it simple and see what we need.
-3. Should there be additional validation to detect and remove other LLM artifacts like explanatory text before or after the JSDoc block? We should ensure we can sanitize this efficiently.
-4. How should the sanitization handle edge cases where the LLM returns multiple JSDoc blocks in a single response? Use your best judgement.
-5. Should the logging level be configurable to allow developers to see detailed processing steps when needed? No need
-6. What is the expected behavior if the sanitization process fails - should it fall back to the original unsanitized response or use the fallback JSDoc generator? I think we can log and silently fail.
-7. Should the success message shown to users be updated to reflect the improved output quality? No need
-8. Are there any performance considerations for the additional sanitization processing step? No
