--- tasks/prd-final-jsdoc-output-cleanup.md
@@ -1,102 +1,0 @@
-# FINAL PRD: JSDoc Generation Output Cleanup
-
-## 1. Overview
-
-This feature addresses critical output quality issues in the existing JSDoc generation functionality. The current `constellation.testDocstringGeneration` command produces duplicate logging and incorrectly formatted JSDoc comments that contain formatting artifacts from the LLM response. The primary goal is to ensure the debug command produces clean, usable JSDoc output that can be directly copied and used by developers, while eliminating redundant console logging that creates noise in the debug console.
-
-## 2. User Stories
-
-- As a developer using the debug command, I want to see the generated JSDoc printed only once so that my debug console remains clean and easy to read.
-- As a developer testing JSDoc generation, I want the output to be a clean, valid comment block free of markdown or nested comment syntax so that I can copy and use it directly.
-- As a developer debugging the extension, I want to be able to trace the JSDoc generation process without redundant logging so that I can efficiently identify issues.
-
-## 3. Functional Requirements
-
-**FR1. Eliminate Duplicate Logging**
-- **Acceptance Criteria:**
-  - Given the `constellation.testDocstringGeneration` command is executed, when the JSDoc generation completes, then the final JSDoc block is logged to the debug console exactly once
-  - Given the command execution, when reviewing the debug console output, then there are no duplicate JSDoc blocks with different prefixes
-  - Given the logging flow, when tracing through the code, then each unique log message appears only once
-
-**FR2. Sanitize JSDoc Output**
-- **Acceptance Criteria:**
-  - Given the LLM returns a response with markdown code fences (```typescript), when the response is processed, then the code fences are completely removed from the final output
-  - Given the LLM returns nested comment blocks, when the response is sanitized, then only the outer JSDoc structure remains
-  - Given any formatting artifacts in the LLM response, when the sanitization process runs, then the output contains only valid JSDoc syntax
-  - Given the LLM returns multiple JSDoc blocks in a single response, when processed, then only the first valid JSDoc block is returned
-
-**FR3. Validate Output Format**
-- **Acceptance Criteria:**
-  - Given a sanitized JSDoc output, when format validation runs, then it confirms the output starts with `/**` and ends with `*/`
-  - Given the final JSDoc output, when copied from the debug console, then it can be directly pasted into a TypeScript file without syntax errors
-  - Given invalid format detection, when the response cannot be properly sanitized, then the system falls back to the existing fallback JSDoc generation
-
-**FR4. Maintain Existing Functionality**
-- **Acceptance Criteria:**
-  - Given the output cleanup implementation, when the command executes, then all existing error handling mechanisms continue to function
-  - Given API failures or network issues, when the fallback JSDoc generation triggers, then it produces clean output without duplication
-  - Given the changes to logging, when the command completes successfully, then the user still receives the appropriate success notification
-
-## 4. Out of Scope (Non-Goals)
-
-- Modifying the LLM model or provider (OpenRouter)
-- Changing the core prompt engineering or system prompts sent to the AI
-- Implementing automatic insertion of JSDoc into source files
-- Modifying the overall architecture or API integration patterns
-- Adding new JSDoc generation features or capabilities
-- Changing the command name or registration mechanism
-- Modifying the test function used for validation
-- Making logging levels configurable
-
-## 5. Technical Considerations
-
-**Duplicate Logging Resolution:**
-- The primary source of duplicate logging occurs in `LLMService.generateDocstring()` method at lines 147-154 where both the raw API response and final validated JSDoc are logged
-- The method logs "Generated JSDoc:" followed by `apiResponseContent` and then "Final validated JSDoc:" followed by `validatedJSDoc`
-- Resolution involves consolidating these into a single, contextual log statement after validation is complete
-
-**Enhanced Sanitization Pipeline:**
-- The existing `validateAndProcessJSDocResponse()` method in `LLMService.ts` (lines 284-303) provides the foundation for enhanced sanitization
-- The current `fixJSDocFormat()` method (lines 337-354) handles basic JSDoc marker addition but needs extension for markdown artifact removal
-- A new sanitization step should be added before format validation to handle:
-  - Code fence removal (```typescript, ```javascript, ```ts, ``` patterns)
-  - Nested comment block extraction
-  - Explanatory text removal before/after JSDoc blocks
-  - Multiple JSDoc block scenarios (select first valid block)
-
-**Integration Points:**
-- The `validateAndProcessJSDocResponse()` method should be enhanced with a new `sanitizeMarkdownArtifacts()` private method
-- The enhanced sanitization will leverage the existing validation pipeline including `validateJSDocFormat()`, `validateContentAndFallback()`, and `generateFallbackJSDoc()`
-- Error handling will follow the established pattern in the service, logging issues and gracefully falling back to the existing fallback JSDoc generator
-- The extension.ts command handler logging (lines 318-320) will remain unchanged as it provides the single final output after processing
-
-**Performance and Reliability:**
-- Sanitization processing will use regex patterns for efficient markdown artifact detection and removal
-- The sanitization process will fail silently with logging when unable to clean response, falling back to existing validation logic
-- No additional dependencies or external libraries required - leveraging existing JavaScript string processing capabilities
-
-## 6. Success Metrics
-
-- Reduce debug console noise by eliminating duplicate JSDoc logging (target: 50% reduction in redundant log entries)
-- Achieve 100% clean JSDoc output format (no markdown artifacts or nested comments in final output)  
-- Maintain 100% compatibility with existing error handling and fallback mechanisms
-- Zero impact on existing test suite execution in `integration.test.ts`
-
-## 7. Implementation Strategy
-
-**Phase 1: Logging Consolidation**
-- Remove duplicate logging statements in `LLMService.generateDocstring()` method
-- Consolidate to single contextual log statement after validation completion
-- Maintain essential debugging context while eliminating redundancy
-
-**Phase 2: Sanitization Enhancement** 
-- Create new `sanitizeMarkdownArtifacts()` private method in `LLMService`
-- Integrate sanitization step into `validateAndProcessJSDocResponse()` pipeline
-- Implement generic pattern matching for common markdown artifacts
-- Handle multiple JSDoc block scenarios with first-valid-block selection
-
-**Phase 3: Validation and Testing**
-- Ensure all existing acceptance criteria for JSDoc generation continue to pass
-- Validate that sanitized output maintains JSDoc format requirements
-- Test edge cases including malformed responses and API failures
-- Verify no impact on existing error handling and fallback mechanisms
