--- tasks/tasks-jsdoc-output-cleanup.md
@@ -1,65 +1,0 @@
-# Task List: JSDoc Generation Output Cleanup
-
-## Relevant Files
-
-- `src/services/LLMService.ts` - The main service class containing JSDoc generation logic that needs duplicate logging removal and sanitization enhancement.
-- `src/test/suite/integration.test.ts` - Existing test suite that needs additional test cases for sanitization functionality.
-- `src/test/unit/LLMService.test.ts` - New unit test file for isolated testing of LLMService sanitization methods.
-
-### Notes
-
-- All changes should maintain backward compatibility with existing error handling and fallback mechanisms.
-- Run tests for the entire test suite with `npm test` or individual test files with `npx mocha [path/to/test/file]`.
-- The implementation follows the existing patterns in LLMService for consistency.
-- Focus on regex-based sanitization for performance and simplicity.
-
-## Tasks
-
-- [ ] 1.0 Eliminate Duplicate Logging in generateDocstring Method
-  - [ ] 1.1 Navigate to `src/services/LLMService.ts` and locate the `generateDocstring()` method around line 147-154.
-  - [ ] 1.2 Remove the duplicate logging statement `console.log('💬 LLMService: Generated JSDoc:', apiResponseContent);` at line 146.
-  - [ ] 1.3 Remove the duplicate logging statement `console.log('📝 LLMService: Final validated JSDoc:', validatedJSDoc);` at line 154.
-  - [ ] 1.4 Add a single consolidated logging statement after validation completion: `console.log('📝 LLMService: JSDoc generation completed:', validatedJSDoc);`.
-  - [ ] 1.5 Verify that all other essential logging statements remain intact for debugging context.
-
-- [ ] 2.0 Create Sanitization Method for Markdown Artifacts
-  - [ ] 2.1 Add a new private method `sanitizeMarkdownArtifacts(content: string): string` in the `LLMService` class.
-  - [ ] 2.2 Implement regex pattern to remove code fences: `/```[\w]*\n?|```\n?/g` for patterns like `\`\`\`typescript`, `\`\`\`javascript`, `\`\`\`ts`, and standalone `\`\`\``.
-  - [ ] 2.3 Implement regex pattern to extract first JSDoc block from multiple blocks: `/\/\*\*[\s\S]*?\*\//` to find the first valid JSDoc comment.
-  - [ ] 2.4 Implement logic to remove explanatory text before and after JSDoc blocks by extracting only the JSDoc content.
-  - [ ] 2.5 Add regex pattern to remove nested comment syntax that might interfere with JSDoc structure.
-  - [ ] 2.6 Add comprehensive logging to the sanitization method following existing LLMService patterns: log input length, artifacts detected, and output preview.
-  - [ ] 2.7 Include error handling that logs issues and returns the original content if sanitization fails.
-
-- [ ] 3.0 Integrate Sanitization into Validation Pipeline
-  - [ ] 3.1 Locate the `validateAndProcessJSDocResponse()` method in `LLMService.ts` around line 284.
-  - [ ] 3.2 Add a call to `sanitizeMarkdownArtifacts()` immediately after the initial `trim()` operation and before format validation.
-  - [ ] 3.3 Update the method to use the sanitized content for all subsequent validation steps.
-  - [ ] 3.4 Add logging to indicate when sanitization is performed: `console.log('🧹 LLMService: Sanitizing markdown artifacts...');`.
-  - [ ] 3.5 Ensure the sanitized content flows correctly through the existing `validateJSDocFormat()` and `validateContentAndFallback()` methods.
-
-- [ ] 4.0 Enhance fixJSDocFormat Method for Edge Cases
-  - [ ] 4.1 Locate the existing `fixJSDocFormat()` method around line 337-354 in `LLMService.ts`.
-  - [ ] 4.2 Add handling for cases where content might have markdown remnants that weren't caught by sanitization.
-  - [ ] 4.3 Ensure the method can handle content that already has partial JSDoc formatting mixed with markdown artifacts.
-  - [ ] 4.4 Verify that the method maintains its existing functionality while being more robust to various input formats.
-
-- [ ] 5.0 Validate Logging Consolidation
-  - [ ] 5.1 Run the `constellation.testDocstringGeneration` command manually and verify debug console output.
-  - [ ] 5.2 Confirm that only one final JSDoc block is logged instead of the previous duplicate logging.
-  - [ ] 5.3 Verify that essential debugging context is maintained while eliminating redundancy.
-  - [ ] 5.4 Test with various scenarios including API success, API failure, and fallback generation to ensure consistent logging behavior.
-
-- [ ] 6.0 Performance and Error Handling Validation
-  - [ ] 6.1 Test sanitization performance with various input sizes to ensure no significant performance impact.
-  - [ ] 6.2 Verify that sanitization failures log appropriately and fall back to original content.
-  - [ ] 6.3 Ensure that all existing error handling paths (API failures, network issues, invalid responses) continue to work correctly.
-  - [ ] 6.4 Test edge cases like empty responses, very large responses, and responses with unusual formatting.
-  - [ ] 6.5 Confirm that the existing fallback JSDoc generation produces clean output without duplication.
-
-- [ ] 7.0 Documentation and Code Review Preparation  
-  - [ ] 7.1 Add JSDoc comments to the new `sanitizeMarkdownArtifacts()` method following existing code patterns.
-  - [ ] 7.2 Update any inline comments in modified methods to reflect the new sanitization step.
-  - [ ] 7.3 Verify that all TypeScript types and interfaces are correctly maintained.
-  - [ ] 7.4 Run the full test suite to ensure no regressions: `npm test`.
-  - [ ] 7.5 Perform manual testing of the `constellation.testDocstringGeneration` command to validate end-to-end functionality.
